refactor(accordion): destructure suppressHydrationWarning on hidden elements
- Replace useHydrationContext() with destructuring suppressHydrationWarning
  from props and spreading it explicitly on hidden elements.
- Move getReactProps test helper to packages/react/test/utils.ts so it can
  be reused across component tests.
